requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
flower==2.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2